Update man page copyright format to match other copyrights
I believe that we took the format of these originally from the gcc
man pages and/or maybe this matched the 1-liner copyright notice
that we used to use.  I don't think there's any reason to preserve
it though, so am updating it to use what we're using elsewhere now.