Deprecate --legacy-classes
Switch the compiler flag `--legacy-classes` to no-op
and the corresponding predefined param `chpl_legacyClasses` to always-false.

Using either of these now causes a deprecation warning.
The intention is to remove them both after v1.21.

Remove all references to `fLegacyClasses` and `gLegacyClasses`
in the compiler and `chpl_legacyClasses` in the modules,
except as needed for the deprecation warnings.

Add tests of the deprecation warnings.