Fix for issue with explicit naming involving nested modules
Turns out that even though we moved resolveModuleCall to before
resolveUnresolvedSymExpr, we could still encounter SymExprs instead of
UnresolvedSymExprs and still needed to handle them.  This only happened in the
case of explicit naming with nested modules.

Note that if we change explicit naming when symbols were renamed to allow using
the new name instead of the old, we will not be able to use the new name in the
error message.  However, that change might not even be implemented, so I'm
punting on it for now.