Allow LIBFABRIC_DIR to override pkg-config, when the latter isn't there.
Some platforms don't have pkg-config.  For such cases, allow overriding
the need for that by means of a LIBFABRIC_DIR environment variable.

(A better solution would be to check pkg-config availability first, and
only look at LIBFABRIC_DIR as a fallback if that isn't there.  But doing
that looked like more of a divergence than I wanted to subject myself to
right now.)