Add a gasnet-smp docker image
Replace the gasnet-experimental image with gasnet-smp. They are similar,
but the smp one uses newer features for oversubscription as well as the
smp substrate (which defaults to segment fast) instead of udp with a
manually set segment fast and local spawning.

The motivation for this is the same as 12903. We want docker images to
do CI testing for Chapel projects, but gasnet udp defaults to segment
everything. udp-everything has more overhead for local testing because
comm will go through the sockets API instead of being short-circuited
with shared memory. This isn't a problem for true multi-locale testing,
but it does impact local oversubscribed testing.