Update Map.map so it works with owned classes
Added a new accessor `this()` function that returns by `const ref` and
added where clauses to it and the const version.

Updated the `init=()` function to take its argument by `ref` instead of
`const ref` because it needs to be able to set the value to `nil` for owned.

Added a simple test that didn't compile previously.