Add unordered GET/PUT support in comm=ofi.
The ofi comm layer has just been doing regular ordered operations for
the unordered GET/PUT/AMO interface calls.  This is functionally correct
but obviously doesn't produce the expected performance benefit.  Here,
really do unordered operations for GET and PUT.  The code was more or
less just copied from the ugni comm layer, changed where necessary for
things like memory registration details.  This should facilitate a
future change in which we move the bulk of the support into common code
and just leave the low-level details of actually initiating and tracking
transactions in the comm layers.

Related to this, the transaction tracking count in the transmit context
descriptors is now overloaded to count just transactions in flight if
the tx context uses a completion queue for tracking, or all transactions
ever initiated if it uses a counter for completion tracking.  This grows
out of how the two tracking mechanisms work: a completion queue reports
each individual transaction as it completes, while a counter reports the
total number of completed transactions so far.  So to tell if everything
initiated has completed, it seems more natural to count the transactions
differently for the two cases.

Unordered AMOs will be dealt with in a separate PR.