Use HaltWrappers in List and Map modules (#14737)
Use HaltWrappers in List and Map modules (#14737)

This PR replaces calls to `halt` in the List and Map modules with
calls to `boundsCheckHalt`, a routine that is part of the HaltWrappers
module. It also sorts the use statements in these modules so that
they are in alphabetical order.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @ronawho for signing off on this!