Try to stabilize syncsingle comm count test
performance/sungeun/multilocale/syncsingle had a variable number of fast-ons
depending on when the begin statement actually completed because the down end
count manipulation could finish before or after we gathered comm counts. This
adds a sync statement to ensure the begin is completed before we gather comm
counts. This lets us remove the prediff that accepted either 1 or 2 fast-ons.

We also saw some tests failures where there were 3 fast-ons. I'm not positive
where the 3rd was coming from, but I was able to reproduce 1/100 times. With
this 5,000 trials passed.