Propagate user thrown errors out of IO routines, channels no longer store error codes (#14739)
Propagate user thrown errors out of IO routines, channels no longer
store error codes (#14739)

This PR wraps up the overhaul of error handling for Chapel IO. It
adjusts IO routines so that they correctly propagate all errors
out of `readThis` and `writeThis` routines.

Prior to this effort, when a user created error was thrown by a
`readThis` or `writeThis` call, it would be caught by IO machinery
and an EIO error code would be stored in the channel as a
placeholder.

In the new world, errors are not stored in channels at all. A user
thrown error will now propagate all the way out of top level IO
routines when thrown.

Code that relied on calls to `channel.error` in order to determine
if an error occurred must now rely on catching the equivalent
Error subclass instead.

Instead of...

  reader <~> new ioLiteral("some prefix");
  if f.error() == EFORMAT {
    // We didn't find the prefix, so do something about it.
    ...
  } else {
    // We found the prefix.
    ...
  }

One must now write:

  try {
    reader <~> new ioLiteral("some prefix");
    // We found the prefix.
    ...
  } catch err: BadFormatError {
    // We didn't find the prefix, so do something about it.
    ...
  }

These changes make it possible for users to throw more descriptive
errors when overloading `readThis` and `writeThis`.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @mppf for review!