Adjust two IO futures (#14832)
Adjust two IO futures (#14832)

Adjust two IO futures that had their failure modes altered in the
wake of #14739. The future `writeThisUserError.chpl` now passes, so
I've changed it into a regular test. The other future still fails,
but the contents of the thrown error have changed.

---

Change to tests, trivial and not reviewed.