Adapt to lack of FI_DELIVERY_COMPLETE with verbs;ofi_rxm provider.
It turns out that the ofi_rxm utility provider, although it advertises
the FI_DELIVERY_COMPLETE capability, only delivers FI_TRANSMIT_COMPLETE
and furthermore, does so silently.  See a conversation on the libfabric
mailing list (libfabric-users@lists.openfabrics.org) in the last few
days for details.  Here, adapt to this by applying a workaround gotten
from the Cray/HPE MPI group, to wit: assert read-after-write ordering,
and then after each batch of unordered PUTs from a given endpoint, do a
GET on the same endpoint from every node that batch did a PUT to.  When
the GETs are complete, we know the PUTs have actually been delivered.

This required the addition of a fair bit of support code, most notably
to track which nodes we've done PUTs to.  Against future needs, for this
I added a generalized bitmap type and operations upon it.

I also added a little dummy heap variable to use for the utility GETs;
we don't need the values there so everyone can share it, but we do need
to be assured we can do RDMA directly to/from it.  I suspect we may find
more uses for this in the future as well.  This can't be the final word
in ordering and completion workarounds ...