Support CHPL_LAUNCHER_REAL_WRAPPER for slurm-gasnetrun* launchers
This is required for the no-op/startup performance test for Cray-CS testing
after 14838. That test wants to time how long the real binary runs for and on
systems with a job scheduler, CHPL_LAUNCHER_REAL_WRAPPER is the way to do that.
We were actually timing the launcher before and just didn't realize it, so this
will also give us a more accurate time.