Add basic parser support for `import Foo;` when Foo is a module
Right now, just pretends that you wrote a `use Foo only;` statement instead,
but the syntax is supported, permits single modules and doesn't permit nested
module access.  I'd like to make an ImportStmt class before fully declaring that
it is working, but this is a great step in the right direction.