We don't necessarily need FI_MULTI_RECV for _all_ receives.
We have been asserting FI_MULTI_RECV in the hints->rx_attr->op_flags
input to fi_getinfo().  This basically says we want that flag on every
receive operation we do.  That's not necessarily so.  Although it is
true that we throw it on all our fi_recvmsg() calls now, we may not do
so forever, and everything works fine without it being in the hints.
Therefore, remove it.

(I just happened to run across this while working on something else.)