Use a different primitive for also resolving body
Resolves a problem with
  test/classes/initializers/compilerGenerated/omittedLoopExpr.chpl
where a toLeader call was calling primitive "call resolves". I suspect
this has to do with an order-of-resolution issue but I decided upon
this workaroud in any case.

Now primitive "call resolves" does not attempt to resolve the fn
called as well, we have "call and fn resolves" for that.