More-principled provider handling.
Our provider handling has been a bit slapdash, and with work on
unordered atomics about to start this is a good time to clean it up.  So
here, add a provider set type and operations upon it, then rearrange and
rewrite so as to do checks against available and in-use providers in the
same way.

Then, get rid of network-based checks.  These weren't necessarily
accurate anyway and are better done as provider-based checks.  For
example, the part of the check in init_fixedHeap() for needing a fixed
heap when using the gni provider on a Cray XC system was effectively
ignoring what provider would be used, and we don't need a fixed heap if
we're using the tcp provider.

While here, improve the getinfo-related debug support.