Propagate character set env vars in chpl_launchcmd
This ensures that the character set environment we've set up on the
login node is correctly propagated to the compute node.

This will resolve failures we're seeing on an XE. We set LC_ALL and
friends in start_test, but that was getting overridden by a login script
that is sourced when the qsub session starts.