Adjust channel.readf overload to throw EFORMAT errors (#14848)
Adjust channel.readf overload to throw EFORMAT errors (#14848)

This PR adjusts an overload of channel.readf so that it does not
return false when an EFORMAT error is encountered. It adjusts the
docstring for channel.readf so that it only returns false when the
channel hits EEOF.

Resolves #14816.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @mppf for review!