Add unordered AMO support in comm=ofi.
The ofi comm layer has just been doing regular ordered operations for
the unordered AMO interface calls.  This is functionally correct but
obviously doesn't produce the expected performance benefit.  Here,
really do unordered operations.  The code was more or less just copied
from the ugni comm layer, changed where necessary for things like memory
registration details.  This follows directly in line with a previous
change that added support for unordered GET and PUT.

Note that this doesn't improve performance on InfiniBand-based systems,
for reasons discussed in the unordered GET/PUT PR (#14810).

While here I also slightly adjusted task_local_buff_acquire(), both the
interface and logic, to allow doing buffer type-specific things outside
the call.  First, there's an additional argument to allow adding some
extra space on the end of the allocated descriptor (for the PUT node
bitmap) and a 'new' indicator in the descriptor so that the caller can
know when acquire() has returned a brand-new descriptor, and do
initialization peculiar to the buffer type.  This allows removing the
type-specific logic that used to be in acquire() itself.