Delay expanding CHPL_HOME.
The common-cray-cs.bash script was evaluating CHPL_HOME at the time it
set CHPL_TEST_LAUNCHCMD.  That's not right.  We're not assured that
CHPL_HOME is set yet, when this script is run.  (Among the things it
does is to set CHPL_HOST_PLATFORM, something that's typically done
before evaluating a setchplenv script which would set CHPL_HOME.)  So
here, escape the dollar sign so that CHPL_HOME will be expanded later,
when CHPL_TEST_LAUNCHCMD is used.