Prevent associative arrays and maps of non-nilable classes (#14846)
Prevent associative arrays and maps of non-nilable classes (#14846)

This PR makes the initialization of associative arrays and maps of
non-nilable classes a compile-time error. For associative arrays,
this is accomplished by adding a `postinit` to the associative
array subclass. For map, this is accomplished via use of a type
routine.

Associative arrays present a complication when the value type of
the array is a non-nilable class. This is because a new key may be
added to the domain of the associative array, causing a nil value to
be default initialized. Additional constraints will have to be
adopted for associative arrays of non-nilable classes to prevent
this from happening.

As for map, it should be possible to enable maps of non-nilable
classes with some effort, see #14861.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @daviditen for review!