Enable '?' on array types
This PR fixes the generated code for 'var data: [dom] (elmType?);'
where 'elmType' is an array type.

When preFold 'elmType?', the compiler used to drop the propagaion
of the runtime type, so uninitialized contents of the runtime-type temps
used to be dropped. This is fixed now.

Explanation for the call to chpl_type_id(): consider the case I came across
when I needed to add the `?` in the field `data` in DefaultAssociativeArr.

Without a call to chpl_type_id, the compiler would generate the node
`move temp2, temp1`, where `temp1` stores `elmType`. This would occur in
wrappers.cpp / buildDefaultedActualFn() while calling DefaultAssociativeArr's
initializer. Later in buildDefaultedActualFn() it calls
`update_symbols(block, &copyMap);` where `copyMap` replaces the above `temp1`
with a (const)ref-intent formal that has a ref type. Therefore the `move`
becomes ill-formed. I tried inserting a PRIM_DEREF into the above `move`,
however could not make that work without unreasonable hackery, either.
Therefore I added a function call to the type identity function, which can
handle both a ref-typed and a non-ref-typed actual.

This issue is exhibited by `test1(ArrType)` in the newly-added test,
`test/classes/nilability/question-mark-on-array.chpl`.

While there, added a --verify compopts to a future test, to highlight
the problem.