Update a few more tests on behalf of comm=ofi.
For launchers that are initiated via system(3), flush stdout before
calling that.  If we don't do this, the verbose output showing the
launcher command line can be buffered during the shell execution and
printed after the output from the user program instead of before it.

In the 'testVFlag' and 'testVerboseFlag' tests, add support for the
slurm-srun launcher and fix flaws on Cray CS.

In the 'args' test, account for a '--walltime' option added by
sub_test.