Report an error upon '?' on a non-class type
Given that we are considering to extend `?` to mean "option" types,
ex. `int?` means "an integer or none", disallow such use for now.
If needed ex. for generic programming, the users can use a type function
such as toNonNilableIfClassType() that I added to
`test/compflags/sungeun/configs/type_variables/configTypeFun.chpl` .

While there, tie a .future test to a GitHub issue and add a --verify compopts
to another .future test, to highlight the problem.