Rely on the VisibilityStmt implementations
Instead of defining versions of checkIfModuleNameMatches on UseStmt and
ImportStmt that just call the parent version, make the parent version not
virtual and remove the child implementations.  Also have ImportStmt rely on the
parent version of isARename() and getRename() - UseStmt cannot due to having
versions with arguments (for when the module's symbols are renamed as part of
the only list).