Change name of methods to avoid issue with overloading parent methods
Prior to this change, UseStmt had to have pretty boring overrides for getRename
and isARename, due to having versions of the same methods defined using
arguments (to enable access to symbols that were renamed in the UseStmt's only
list).  This changes the name of those methods so that they no longer conflict
with the argumentless versions on VisibilityStmt, and removes the dummy
overrides.