Cleanups.
In the process of working on the default-completion change I ran into
some flawed code, so I'm taking the opportunity to fix that here.
First, the transaction tracking type code would have required 2 bits of
storage if we actually used its highest value, but we only had allocated
1 bit for it in the encoding.  Fortunately we didn't use that highest
value, so no harm was done.  To resolve this now and in the future, do a
better job of parameterizing the encoding and also add some debug-only
checks to make sure everything fits.

Also, add a missing 'destroy' call to match the 'init' for an atomic
'transaction done' indicator.

While here I also deleted an unused function and fixed a white space
flaw (picky, picky).