Wait longer for unordered PUTs to finish.
Testing of some other changes seemed to indicate a sporadic race in
unordered PUT handling once the delivery_complete requirement was
relaxed.  Specifically, we would sometimes proceed past an unordered-PUT
fence before all the the PUT values were actually visible.  The existing
code relied on the fact that, with read-after-write ordering and
delivery_complete, firing a GET to all the nodes we'd PUT data to would
push all the PUT data into target memory before the fi_write() returned.
But with transmit_complete that seems not to be the case, necessarily.
So here, not only fire such a GET, but also wait for its completion to
arrive.  This seems to resolve the PUT races we observed.