Implement the new compareExchange API for processor atomics
The new compareExchange version takes `expected` by ref and updates it
on failure. This matches the C/C++ API. The implementation is pretty
straightforward. The main thing is that we have to ensure that we
localize our expected value when we go to do operations.

The runtime implementations for `cstdlib` and `locks` are pretty simple.
The `intrinsics` version is a little bit trickier since we have to
implement compareExchange with a compareAndSwap primitive that only
returns the old value. This implementation has been inspired by:
https://herbsutter.com/2014/02/19/reader-qa-is-stdatomic_compare_exchange_-implementable/