Implement the new compareExchange API for network atomics
We again have to localize `expected`, but in this case we have to make
sure it's localized on the initiating node, not the target node. ugni
and ofi only have a compareAndSwap that returns the old value available
for the NIC/FI_ implementations. Update our processor atomics fallbacks
in ugni/ofi to behave similarly, and compute success/failure and update
`expected` at the callsite.