Rename ugni AMO opcodes from xchg to swap
Compare-and-swap and compare-exchange are largely interchangeable terms
(in the processor and language communities), but at the Chapel interface
level they mean different things. Here rename the internal AMO/NIC op
codes to reflect that our NIC and processor atomics are doing a
traditional CAS, where the old value is returned. We then use that to
implement the Chapel compareExchange interface.