Use localAccess for ra-on
Switch from using a local block to localAccess for ra-on. This allows
the fast-on optimization to trigger, which eliminates task creation. The
local block was originally added in 4a8cedf8fc to enable the fast-on
optimization, but it looks like there's been a regression since then.
With the local block today we're still emitting PUT/GET calls (that end
up being local) that thwart the fast-on optimization. Longer term we
want to deprecate support for the local block anyways, so switch to
localAccess.

RA-on performance for XC (Aries/ugni) and CS (FDR-IB/gasnet-ibv-large):

| config  | Before      | After       |
| ------- | ----------- |  ---------- |
| Cray-XC | 0.0119 GUPS | 0.0143 GUPS |
| Cray-CS | 0.0018 GUPS | 0.0027 GUPS |