Remove mention of old Qthreads-resident Chapel tasking shim library.
Chapel's Qthreads-based tasking shim originally resided within the
QThreads repo rather than the Chapel runtime, and was built into
libqthread_chpl.a.  But the shim moved into the runtime proper long ago
and that library hasn't been built in years.  Here, remove the last
remnant mention of it, along with the `-lchpl` that needed to follow it
on the user program link line.  And since the only other thing in the
`default_get_link_args(libs=)` argument is what that function searches
for by default, remove that `libs` argument altogether.