Use getVisibleMethods strategy for certain function names
Things like _new, _defaultOf, and operators should be treated like methods on
the type.  In the case of the first two, this is because they are essential for
its operation.  In the case of the last, this is because there is no way for the
user to specify whether they should be included or excluded on only/except lists

Note that due to a bug in resolution, this does not actually fix #14740 (user
defined overload of class assignment not detected when use is outside defining
module).