Drop back from delivery_complete to the provider's default level.
We have been specifying the FI_DELIVERY_COMPLETE completion level for
our reads, writes, atomics, and sends (AMs).  Here, drop back to the
provider's default level instead.  This is still FI_DELIVERY_COMPLETE,
effectively, for reads and fetching atomics.  That is, the completion is
event is not delivered until the requested data lands in local memory,
as documented in the fi_cq(3) man page.  But for writes, non-fetching
atomics, and sends it is whatever the provider believes will perform
best while still delivering the required libfabric semantics.

We're doing this for two reasons: it should or at least may perform
better, and at least one provider (the ofi_rxm utility provider used
with a variety of base providers including verbs) actually cannot do
delivery_complete and silently provides transmit_complete instead.  This
latter bug is known, but is not likely to to be fixed soon, so we have
to work around it.

Obviously we still have to adhere to the Chapel Memory Consistency Model
(MCM) even though as of this change we no longer know the relationship
between when we see the completion for a write or a send and when the
effects of that write or send are observable.  We do so by requesting
read-after-write and send-after-write transaction ordering and then
doing reads to nodes in order to effectively force release consistency
on our writes to those same nodes.  Task-private bitmaps track which
nodes each task has written to since its last such release "fence".

There are several related changes here also.  The biggest one is the
addition of chpl_comm_release(), which does the release-forcing reads
described above when it is called from chpl_rmem_consist_release()
before atomic operations and on-stmts.  (This happens through the module
code and compiler.)  Releases at other times are initiated from inside
the comm layer itself.

Also in related work, I needed to augment the bitmap type somewhat,
mainly by adding the 'alloc' and 'free' routines I had earlier thought
might not be needed.

This current solution has one aspect that is overkill.  The MCM requires
that within a task, regular loads and stores to the same address must be
seen to occur in program order.  The current code enforces this by doing
a write release (by means of a dummy read, as above) to the target node,
before initiating a read from that node.  The addresses are not taken
into account, only the node.  In the future we expect to do something
smarter for this case, perhaps implementing a write cache and satisfying
subsequent reads out of that.  There's a TODO comment in the code to
this effect.