Only mark processor atomics as safe for fast-ons
In 4008, we had the compiler mark extern functions in the atomic modules
as safe for fast-ons so that downEndCount could run in the AM-handler
for network-atomics=none. The notion was that processor atomic calls are
safe for fast-ons and that is true, but previously this compiler code
also considered network atomics as safe for fast-ons, which isn't true
since network atomics may use arbitrary comm.

Instead of having the compiler automatically mark functions that are
safe, just manually mark them in the processor atomic module code. Note
that previously we didn't consider locks-atomics as safe for fast on,
but now that we're using pthread mutexes for them it should be ok to run
them in the am handler. It may limit the rate of AMs for atomics=locks
in some cases, but that's not an important configuration anyways.