Fix RA comm counts
Update ra_commDiags for 15001. 15001 updated RA to use localAccess and
switched `on TableDist.idxToLocale[r&indexMask]` to `on T[r&indexMask]`.
Switching the on clause added some GETs from populating the RADCache,
but we believe this code is more elegant and the number of GETs per
locale is constant.

Testing this on XC revealed that verification was failing. Update the
table size to reduce conflicts like we did in 36c521bfab for similar
reasons and remove non-LCG version originally added in 8d68e2cdc0, so
that we're testing the same config as release/examples.