Change expected behavior and contents of this test to reflect consensus
Now that we've decided that importing a module using the full path to it will
make only the most recent module's symbols available using just its name as the
qualified name (e.g. `import A.B;` will enable `B.x`), update the test to
utilize this strategy to refer to B's symbols.  Also change the expected output
in anticipation of implementing it.