Use (synthetic) delivery_complete semantics for ordered PUTs, for now.
The completion level reduction in this PR was written to rely on the
existing remote memory consistency support for the required release
fences, but it turns out that's only enabled with remote caching.  We're
not confident enough in remote caching to turn it on by default so close
to release, so here, just switch to using synthesized delivery_complete
semantics for ordered PUTs.

As part of this, the task-private nodeWriteInfo data is removed, along
with everything that generated or used it.