Bump ugni from 4K to 16K memory regions
We have had users in the field run out of memory regions on 256GB nodes.
Assuming 16M hugepages, 16K memory regions is the most that could
possibly be needed on 256GB nodes, so bump our default to be safe.

Each node maintains a list of every other nodes memory regions, so we
use `num_memory_regions * sizeof(memory_region) (32 bytes) * num_nodes`
bytes of memory for memory regions. At 1024 nodes this increases the
per-node memory usage for mem regions from 128MB to 512MB. That's a
non-trivial amount of memory, but realistically 1-2K nodes is probably
the largest that anybody will run Chapel at on an XC.