Avoid more comm/compute overlap in ugni
In 12960 we eliminated comm/compute overlap for FMA operations when a
task hadn't done ~100 FMA operations. Here we extend that to RDMA and
chained FMA operations as well. This is motivated by aggregation work
where we do on-stmts followed by large RDMA PUTs/GETs, but we really
don't want too many tasks started at once since that increses memory
presssure. This also adds an option to completely disable comm/compute
overlap in ugni, but it's not enabled by default since it has a 2x
performance hit for SSCA.