Revert the addition of chpl_comm_release().
A previous commit in this PR added chpl_comm_release() so that we could
use it for release fences to force comm=ofi PUT data to be visible in
target memory.  But with the change to synthetic delivery_complete
semantics for ordered PUTs in comm=ofi (see previous commit) we no
longer need chpl_comm_release(), so back out its addition for now.