Snapshot of changes so far
Switches the parser to utilize exprs instead of just idents for import statement
modules.  Adds a private isValid method to ImportStmts similar to the one
defined on UseStmts, though we expect them to diverge both for error messages
and in supported functionality (hence not making it a common method on
VisibilityStmts).  Utilizes this new method in ImportStmt::scopeResolve.

This does not allow nested modules to be imported just yet, because the error
message for importing them relatively (i.e. in the scope in which they are
defined without using the full path) is applying to them generally.  I need to
implement another strategy to find modules for import statements because of
their different limitations