Make some debugging-related improvements.
Assert that the AM handler is never asked to initiate an AM itself.

Make a distinction between 'read' AMO operations and all others for
debugging purposes.  In certain circumstances reads are used in wait
loops, and being able to look at AMOs without all that traffic can be
useful.

Allow for fi_wait() returning -FI_EINTR, since that seems to happen
fairly commonly if we're executing in a debugger session, ^C the program
to see where it's at, and then continue.