Adjust signature of list.sort to accept multiple comparator types (#14999)
Adjust signature of list.sort to accept multiple comparator
types (#14999)

This PR adjusts the signature of `list.sort` so that it is generic
over one argument, the comparator to use, with the default type
being `Sort.defaultComparator`. It also adds one new test to lock
in this behavior.

Resolves #14915.

---

Testing:

- [x] ALL on linux64 when CHPL_COMM=none
- [x] ALL on linux64 when CHPL_COMM=gasnet

---

Thanks to @e-kayrakli for review!