Fix another misaligned GET bug in ugni
Under ugni, the size and addresses of a GET have to be 4-byte aligned.
If they aren't, we use bounce buffers and over-GET in order to meet
those alignment requirements. We had a bug in this code, where we could
end up requesting too small of a bounce buffer and overflow it,
corrupting the bounce buffer data structures.

For a 7-byte GET with a 7-byte aligned address, we'd end up doing a
16-byte GET, but we were only asking for an 8-byte buffer. We'd overflow
that buffer, corrupting the locks and other data structures around it.
Here we ensure that we ask for a bounce buffer that's at least as large
as the aligned transmission size, not the original size.

It was pretty rare to trigger this bug because you needed the GET size
to be < 8, and the aligned GET size to be > 8. That requires a 2,3,5,6,
or 7 byte GET with wrong alignment. Most of our codes end up doing 1, 4,
or 8 byte GETs, all of which worked correctly.