Change expected behavior of same_name_conflict and add a new test
Imports first resolve to modules brought in by use and import statements in
scope, then looking at those in outer scopes, until finally we reach the top
level module scope, at which point we look at top-level module symbols.  This
means that bringing in a module and renaming it to the same name as a top-level
module shadows the top-level module for the scope with that import.  This
matches Rust's behavior, with the exception of one thing - Rust complains if you
import a module and then try to import it again (which is what this code is now
doing), but we don't do that for use statements.

Add a new test to demonstrate the behavior when renaming a module, and then
using the new name to import a module defined in the renamed module