add 'owned.create(c)' and 'shared.create(c)' (#15119) by vass
Resolves #14300.

Defines `owned.create(c)` and `shared.create(c)` factory methods. They return
an owned/shared pointer to the argument, which must be a class. They take over
or share the ownership of the argument as appropriate. The naming follows
the leading choice in #14291.

The argument types are allowed as follows, to match what was allowed
using the deprecated forms (see below):

when `c` is:     `owned.create(c)` and `shared.create(c)` are:
 owned             allowed only when `c` is nilable or expiring
 shared            only `shared.create(c)` is allowed
 borrowed          not allowed
 unmanaged         allowed
 not a class       not allowed

Deprecates `new owned(c)` and `new shared(c)` in favor of `.create()`.

Adjusts a handful of tests. Replaces `new shared(nil:unmanaged MyClass?)`
with simply `nil` or, when not enough, with `nil: shared MyClass?`.
Adds testing of `.create()`.

IMPLEMENTATION DETAILS

Currently the deprecation warnings are issued in the parser specifically
for the two deprecated forms. If I added warnings to the corresponding
initializers for _owned / _shared, it would also catch the cases like
`type T = owned; ... new T(c) ...`. However, those initializers are also
used for `new owned MyClass()` and `new shared MyClass()`, so I chose
to put them in the parser instead. Once we remove those parser rules
altogether, we can revisit.

I chose to introduce catch-all implementations of `.create()` so that
the compiler can issue clear error messages. The error of invoking
`.create()` on a non-nilable non-expiring `owned` is issued later
by the lifetime checker.

FUTURE DIRECTIONS

We may want to support `type t = shared MyClass?; t.create(nil);`
For that, create an overload `proc shared.create(arg: _nilType)`
and report an error if 'this' is not of a concrete type.

It would be nice to provide the follow library function to replace
several uglier uses of `parseToml` in `tools/mason`:

    proc parseTomlFile(filename: string): owned Toml

r: @mppf